refactor: Data 모듈에서 User 관련 파일들을 별도 폴더로 분리
- Beverage 폴더 내의 User 관련 파일들을 User 폴더로 이동
- Repository, Request, Response, Target 구조 유지
- UserRepository.swift, UserRecordRequest.swift 등 10개 파일 이동
- 코드 구조 개선으로 모듈 책임 분리 강화
