+++ HumanNeverDie/Projects/Data/Data/Sources/User/Request/UserRecordRequest.swift
@@ -21,7 +21,6 @@ struct NotificationsRequest: Encodable {
   let isEnabled: Bool
 }
 
-
 struct UserNotificationsRequest: Encodable {
  let isEnabled: Bool
  let remindersEnabled: Bool
